+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Removed `AssetType` in favor of classes that implement `Asset` directly (e.g. `BankAccountAsset`)
 * Reordered package structure to be more sensible
 * Split deserialization & error-checking for `AssetsHistory` and `AssetDefinitions` to make code easier to manage
+* Drop the `-example` suffix to all the example files
 
 ## 0.6.0
 * Refactor the data model around assets in preparation for being able to define per-asset growth rates